Cleanup NCMenuActions
