+++ iOSClient/Menu/NCMenuCells.swift
@@ -2,14 +2,29 @@
 //  NCMenuCells.swift
 //  Nextcloud
 //
-//  Created by Henrik Storch on 10.12.21.
-//  Copyright © 2021 Marino Faggiana. All rights reserved.
+//  Created by Henrik Storch on 13.12.2021.
+//  Copyright © 2021 Henrik Storch. All rights reserved.
+//
+//  Author Henrik Storch <henrik.storch@nextcloud.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //
 
 import Foundation
 import UIKit
 
-class NCMenuTextFIeldCell: UITableViewCell, NCMenuCell, UITextFieldDelegate {
+class NCMenuTextFieldCell: UITableViewCell, NCMenuCell, UITextFieldDelegate {
     @IBOutlet weak var label: UILabel!
     @IBOutlet weak var icon: UIImageView!
     @IBOutlet weak var textField: UITextField!
@@ -109,17 +124,3 @@ extension NCMenuCell {
     }
 }
 
-extension UIView {
-    // https://stackoverflow.com/a/24590678
-    var parentViewController: UIViewController? {
-        // Starts from next (As we know self is not a UIViewController).
-        var parentResponder: UIResponder? = self.next
-        while parentResponder != nil {
-            if let viewController = parentResponder as? UIViewController {
-                return viewController
-            }
-            parentResponder = parentResponder?.next
-        }
-        return nil
-    }
-}
